+++ app/src/main/java/com/example/shoppinglist/features/listDetailScreen/domain/entity/Product.kt
@@ -1,8 +1,8 @@
 package com.example.shoppinglist.features.listDetailScreen.domain.entity
 
 data class Product(
-    val id: Int = 0,
-    val listId: Int,
+    val id: Long = 0,
+    val listId: Long,
     val name: String,
     val quantity: Double = 0.0,
     val unit: String? = null,
