+++ app/src/main/java/com/example/shoppinglist/features/listDetailScreen/domain/repository/ProductsRepository.kt
@@ -4,11 +4,11 @@ import com.example.shoppinglist.features.listDetailScreen.domain.entity.Product
 import kotlinx.coroutines.flow.Flow
 
 interface ProductsRepository {
-    fun getAllByListId(listId: Int): Flow<List<Product>>
-    suspend fun deleteById(id: Int)
-    suspend fun deleteAllByListId(listId: Int)
-    suspend fun deleteCheckedByListId(listId: Int)
+    fun getAllByListId(listId: Long): Flow<List<Product>>
+    suspend fun deleteById(id: Long)
+    suspend fun deleteAllByListId(listId: Long)
+    suspend fun deleteCheckedByListId(listId: Long)
     suspend fun update(product: Product)
     suspend fun create(product: Product)
-    suspend fun toggleChecked(id: Int)
+    suspend fun toggleChecked(id: Long)
 }
