+++ app/src/main/java/com/example/shoppinglist/features/listDetailScreen/domain/usecase/DeleteProductByIdUseCase.kt
@@ -6,7 +6,7 @@ import jakarta.inject.Inject
 class DeleteProductByIdUseCase @Inject constructor(
     private val repository: ProductsRepository
 ) {
-    suspend operator fun invoke(id: Int) {
+    suspend operator fun invoke(id: Long) {
         repository.deleteById(id)
     }
 }
