+++ app/src/main/java/com/example/shoppinglist/features/listDetailScreen/domain/usecase/GetProductsByListIdUseCase.kt
@@ -8,7 +8,7 @@ import kotlinx.coroutines.flow.Flow
 class GetProductsByListIdUseCase @Inject constructor(
     private val repository: ProductsRepository
 ) {
-    operator fun invoke(listId: Int): Flow<List<Product>> {
+    operator fun invoke(listId: Long): Flow<List<Product>> {
         return repository.getAllByListId(listId)
     }
 }
